requirements.txt
langchain
openai
unstructured
pillow
transformers
torch
bs4
Flask
flask_restful
cohere
pinecone-client
